Address issue #306 - use misalign instead of evbuffer_remove
when evbuffer_add_file is called and mmap is used, if the offset
argument is >0, a mistake happens: add_file removes "offset" byts
from the front of the evbuffer.

So that means any data that was previously on the buffer is trimmed
off by "offset" bytes. Whoops.

A onelinter fix: don't use evbuffer_drain for the offset, instead,
just modify the misalign variable on the newly created chain.